+++ android/src/main/kotlin/com/albertalrisa/flutter/plugins/shareapi/requests/Instagram.kt
@@ -1,3 +1,3 @@
 package com.albertalrisa.flutter.plugins.shareapi.requests
 
-val INSTAGRAM_SHARE_TO_STORY = 301
+const val INSTAGRAM_SHARE_TO_STORY = 5301