+++ android/src/main/kotlin/com/albertalrisa/flutter/plugins/shareapi/requests/SystemUI.kt
@@ -1,0 +1,5 @@
+package com.albertalrisa.flutter.plugins.shareapi.requests
+
+const val SYSTEMUI_SHARE_TEXT = 5101
+const val SYSTEMUI_SHARE_IMAGE = 5102
+const val SYSTEMUI_SHARE_FILE = 5103