+++ lib/share_api.dart
@@ -10,11 +10,6 @@ class ShareApi {
 
   static const MethodChannel _channel = const MethodChannel(channel_name);
 
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
-  }
-
   static final viaSystemUI = SystemUI(_channel);
   static final viaFacebook = Facebook(_channel);
   static final viaInstagram = Instagram(_channel);