+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log (share_api)
 
+## 0.0.3
+
+* **Instagram Share Result Fix and Default Function Removal**
+* Instagram should now return `ShareResult.unknown` instead of `ShareResult.canceled` (breaking change)
+* Built in `platformVersion()` function from Flutter default plugin example has been removed (breaking change)
+* Documentation now includes the return value explanation
+
 ## 0.0.2
 
 * **Documentation Revamp**